+++ R/plot_scseq.R
@@ -328,7 +328,7 @@ downsample_clusters <- function(scseq, max.cells = 200) {
 #' @return list used by \link{VlnPlot}
 #'
 #' @keywords internal
-get_ridge_data <- function(feature, scseq, selected_cluster, by.sample = FALSE, decreasing = feature %in% c('ribo_percent', 'log10_sum', 'log10_detected'), with_all = FALSE, dgrlogs = NULL) {
+get_ridge_data <- function(feature, scseq, selected_cluster, by.sample = FALSE, decreasing = feature %in% c('ribo_percent', 'log10_sum', 'log10_detected'), with_all = FALSE, h5logs = NULL) {
   n <- NULL
 
   if (isTruthy(selected_cluster)) {
@@ -385,8 +385,8 @@ get_ridge_data <- function(feature, scseq, selected_cluster, by.sample = FALSE, 
 
 
   if (feature %in% row.names(scseq)) {
-    if (!is.null(dgrlogs)) {
-      x <- fast_dgr_row(dgrlogs, feature)
+    if (!is.null(h5logs)) {
+      x <- h5logs[feature, ]
       x <- x[colnames(scseq)]
     } else {
       x <- as.numeric(SingleCellExperiment::logcounts(scseq[feature, ]))